backend/foundit-backend-spring/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
ultralytics==8.1.0
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
